requirements.txt
certifi==2021.5.30
cycler==0.11.0
dataclasses==0.8
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
torch==1.10.2
tqdm==4.64.1
typing_extensions==4.1.1
zipp==3.6.0
PyYAML==6.0
